Created BitBuffer module with more methods
